requirements.txt
dash
pandas
numpy
scipy
langgraph
langchain
langchain-openai
plotly
requests
pyarrow
python-dotenv
